opencvBackup/requirements.txt
numpy==1.14.0
opencv-contrib-python==3.4.0.12
opencv-python==3.4.0.12
Pillow==5.0.0
